refactored blockies - removed unused code
