+++ android/src/main/java/io/parity/signer/components/blockies/svalinn/Blockies.kt
@@ -1,9 +1,10 @@
 package io.parity.signer.components.blockies.svalinn
 
 /**
- * Based on svalinn-kotlin project which is MIT licensed.
+ * Based on svalinn-kotlin project.
+ * Reimplemented for Vault use case
  */
-class Blockies(
+internal class Blockies(
     val primaryColor: Int,
     val backgroundColor: Int,
     val spotColor: Int,
@@ -93,5 +94,15 @@ class Blockies(
             val blue = (b * 255).toInt()
             return 0xFF shl 24 or (red shl 16) or (green shl 8) or blue
         }
+
+			private fun hueToRGB(p: Float, q: Float, h: Float): Float {
+				var hue = h
+				if (hue < 0) hue += 1f
+				if (hue > 1) hue -= 1f
+				if (6 * hue < 1) return p + (q - p) * 6f * hue
+				if (2 * hue < 1) return q
+
+				return if (3 * hue < 2) p + (q - p) * 6f * (2.0f / 3.0f - hue) else p
+			}
     }
 }
