--- android/src/main/java/io/parity/signer/components/blockies/svalinn/readme.md
@@ -1,3 +1,0 @@
-This package it based on MIT licensed project svalinn-kotlin,
-It is are modified to use with string seeds and in compose.
-
